requirements.txt
streamlit
google-genai
pandas
Pillow